+++ src/sclab/examples/processor_steps/__init__.py
@@ -1,5 +1,7 @@
 from ._cluster import Cluster
 from ._differential_expression import DifferentialExpression
+from ._gene_expression import GeneExpression
+from ._integration import Integration
 from ._neighbors import Neighbors
 from ._pca import PCA
 from ._preprocess import Preprocess
@@ -10,8 +12,10 @@ __all__ = [
     "QC",
     "Preprocess",
     "PCA",
+    "Integration",
     "Neighbors",
     "UMAP",
     "Cluster",
+    "GeneExpression",
     "DifferentialExpression",
 ]