+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Dataset Integration component
+- GeneExpression visualization component
+
+### Fixed
+- Create only one layer during preprocessing
 
 ## [0.2.3] - 2025-04-03
 ### Dependencies